Dialog Updates
Reworked organizational structure for dialog boxes. Added new InformationDialog popup which contains help pages and browser compatibility information.
