--- web/src/Components/Welcome/Dialog.css
@@ -1,22 +1,0 @@
-.COElogo{
-    padding: 10px;
-    width: 100%;
-    height: auto;
-}
-
-.CSlogo{
-    width: 50%;
-    height: auto;
-}
-
-.NLPlogo{
-    padding: 10px;
-    width: 50%;
-    height: auto;
-}
-
-#acknowledgements-button {
-    position: absolute;
-    bottom: 25px;
-    right: 25px;
-}
